requirements.txt
selenium==4.15.2
pymongo==4.6.0
python-dotenv==1.0.0
flask==3.0.0
webdriver-manager==4.0.2
requests==2.31.0
urllib3<2.0.0
undetected-chromedriver==3.5.4
